- expose length parameter to influence connection length of node connectors in plantuml diagrams - introduce ELK layout in mermaid charts - refactor python structure to be more inspecific about diagram types generated
