+++ src/ili2py/writers/diagram/interlis/templates/dot/class.jinja2
