+++ src/ili2py/writers/diagram/interlis/templates/dot/classes.jinja2
