+++ src/ili2py/writers/diagram/interlis/templates/dot/dot.jinja2
