+++ src/ili2py/writers/diagram/interlis/templates/plantuml/plantuml.jinja2
@@ -16,7 +16,7 @@ namespace {{ model_group.oid }} {
 {% endif %}
 {%- for association in model_group.associations %}
 {%- if len(association) == 3 %}
-{{ association[0][0] }} "{{ association[0][1] }}" {{ association[0][2] }}--{{ association[1][2] }} "{{ association[1][1] }}" {{ association[1][0] }} : {{ association[2] }}
+{{ association[0][0] }} "{{ association[0][1] }}" {{ association[0][2] }}{{ "-" * multiplier }}{{ association[1][2] }} "{{ association[1][1] }}" {{ association[1][0] }} : {{ association[2] }}
 {%- else %}
     note "Association <b>'{{ model_group.name }}.{{ association[-1] }}'</b>\n has more than 2 sides,\n this is currently not supported" as N{{ randint(1,10000) }}
 {%- endif %}
@@ -24,7 +24,7 @@ namespace {{ model_group.oid }} {
 {%- for class_definition in model_group.classes  %}
 {%- if class_definition.inherits_from %}
 {%- if class_definition.inherits_from.split('.')[0] in model_names  %}
-{{ class_definition.oid }} --|> {{ class_definition.inherits_from }}
+{{ class_definition.inherits_from }} <|{{ "-" * multiplier }} {{ class_definition.oid }}
 {%- endif %}
 {%- endif %}
 {% endfor %}
@@ -43,7 +43,7 @@ namespace {{ topic_group.oid }} {
 {%- for association in topic_group.associations %}
 {%- if len(association) == 3 %}
 {%- if index.index[association[0][0]].element_in_package.ref.split('.')[0] in model_names and index.index[association[1][0]].element_in_package.ref.split('.')[0] in model_names %}
-{{ association[0][0] }} "{{ association[0][1] }}" {{ association[0][2] }}--{{ association[1][2] }} "{{ association[1][1] }}" {{ association[1][0] }} : {{ association[2] }}
+{{ association[0][0] }} "{{ association[0][1] }}" {{ association[0][2] }}{{ "-" * multiplier }}{{ association[1][2] }} "{{ association[1][1] }}" {{ association[1][0] }} : {{ association[2] }}
 {%- endif %}
 {%- else %}
     note "Association <b>'{{ model_group.name }}.{{ topic_group.name }}.{{ association[-1] }}'</b>\n has more than 2 sides,\n this is currently not supported" as N{{ randint(1,10000) }}
@@ -52,7 +52,7 @@ namespace {{ topic_group.oid }} {
 {%- for class_definition in topic_group.classes  %}
 {%- if class_definition.inherits_from %}
 {%- if class_definition.inherits_from.split('.')[0] in model_names  %}
-{{ class_definition.oid }} --|> {{ class_definition.inherits_from }}
+{{ class_definition.inherits_from }} <|{{ "-" * multiplier }} {{ class_definition.oid }}
 {%- endif %}
 {%- endif %}
 {% endfor %}
