+++ src/ili2py/writers/diagram/interlis/uml.py
